Viper MFD/SMS Builder Updates
Updated MFD and SMS builders for the Viper. Lua support updated to operate with latest DCS behavior. Now correctly handle the case where MFD configuration is default, but SMS configuration is not. Only program Mav settings once when mixed types in munitions (as configuration is common in this case). Compiler warning fixes and cleanup.